requirements.txt
pandas>=1.3.0
numpy>=1.20.0
streamlit>=1.12.0
matplotlib>=3.4.0
seaborn>=0.11.0
openpyxl>=3.0.0
xlsxwriter>=3.0.0
requests>=2.25.0
pyyaml>=6.0
pytest>=6.0.0
tqdm>=4.62.0
plotly>=5.10.0
